Registration script workings added
I have now added the code to make the registration script work.  At the moment this is only a basic code, which will be changed once the main part of the project is sorted.  

I will be implementing encryption for passwords, filtering for all inputs (username, email address) to ensure no rogue characters are inserted that could compromise the DB